refactor(esl-utils): move `isElement` to `dom/api` module
