+++ src/modules/esl-media/core/esl-media.ts
@@ -1,16 +1,15 @@
-import {isElement} from '../../esl-utils/misc/object';
-import {ExportNs} from '../../esl-utils/environment/export-ns';
 import {ESLBaseElement} from '../../esl-base-element/core';
-import {bind, prop, attr, boolAttr} from '../../esl-utils/decorators';
+import {ExportNs} from '../../esl-utils/environment/export-ns';
+import {isElement} from '../../esl-utils/dom/api';
 import {CSSClassUtils} from '../../esl-utils/dom/class';
-import {rafDecorator} from '../../esl-utils/async/raf';
-import {debounce} from '../../esl-utils/async/debounce';
+import {SPACE, PAUSE} from '../../esl-utils/dom/keys';
+import {bind, prop, attr, boolAttr} from '../../esl-utils/decorators';
+import {debounce, rafDecorator} from '../../esl-utils/async';
 import {parseAspectRatio} from '../../esl-utils/misc/format';
 
 import {ESLMediaQuery} from '../../esl-media-query/core';
 import {ESLTraversingQuery} from '../../esl-traversing-query/core';
 
-import {SPACE, PAUSE} from '../../esl-utils/dom/keys';
 import {getIObserver} from './esl-media-iobserver';
 import {PlayerStates} from './esl-media-provider';
 import {ESLMediaProviderRegistry} from './esl-media-registry';
